pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neural-net-encryption"
version = "0.1.0"
description = "Experimental neural cryptography research lab"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "numpy>=1.26",
  "PyYAML>=6.0",
  "matplotlib>=3.8",
  "cryptography>=42.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
ml = ["torch>=2.5"]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["error"]
